refactor(monitor): move MonitorSpherical to monitors.py
